fix : 화폐단위 변경하면 SearchBar 검색에 적용안되는 오류 수정을 위한 SearchBar 별도의 Scene으로 분리
