+++ ReceiptManager/ReceiptManager/Source/Scene/Scene.swift
@@ -16,6 +16,7 @@ enum Scene {
     case large(LargeImageViewModel)
     case selectImage(SelectImageViewModel)
     case setting(SettingViewModel)
+    case search(SearchViewModel)
 }
 
 extension Scene {
@@ -59,6 +60,11 @@ extension Scene {
         case .setting(let settingViewModel):
             let viewController = SettingViewController()
             viewController.bind(viewModel: settingViewModel)
+            
+            return viewController
+        case .search(let searchViewModel):
+            let viewController = SearchViewController()
+            viewController.bind(viewModel: searchViewModel)
             
             return viewController
         }
