--- ReceiptManager/ReceiptManager/Source/View/MainScene/SearchView.swift
@@ -1,40 +1,0 @@
-//
-//  SearchView.swift
-//  ReceiptManager
-//
-//  Created by parkhyo on 2023/06/08.
-//
-
-import UIKit
-
-final class SearchView: UIView {
-    let tableView = UITableView(frame: .zero, style: .insetGrouped)
-    
-    override init(frame: CGRect) {
-        super.init(frame: frame)
-        setupView()
-        setupConstraint()
-    }
-    
-    required init?(coder: NSCoder) {
-        fatalError("init(coder:) has not been implemented")
-    }
-    
-    private func setupView() {
-        backgroundColor = ConstantColor.backGrouncColor
-        addSubview(tableView)
-        
-        tableView.backgroundColor = ConstantColor.backGrouncColor
-        tableView.register(ListTableViewCell.self, forCellReuseIdentifier: ListTableViewCell.identifier)
-        tableView.translatesAutoresizingMaskIntoConstraints = false
-    }
-    
-    private func setupConstraint() {
-        NSLayoutConstraint.activate([
-            tableView.topAnchor.constraint(equalTo: topAnchor),
-            tableView.leadingAnchor.constraint(equalTo: leadingAnchor),
-            tableView.trailingAnchor.constraint(equalTo: trailingAnchor),
-            tableView.bottomAnchor.constraint(equalTo: bottomAnchor)
-        ])
-    }
-}
